pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "willian", email = "nobrewillian756@gmail.com"},
    {name = "williannobree", email = "nobrewillian756@gmail.com"},
]
dependencies = [
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false
